+++ src/utils/dataHelper.js
@@ -69,3 +69,23 @@ export function setScenesForCartLayer() {
   }
   addDataToLayer(cartGeojson, 'cartFootprintsLayer', options)
 }
+
+export function processDisplayFieldValues(value) {
+  if (typeof value === 'boolean') {
+    return value.toString()
+  } else if (Array.isArray(value)) {
+    return value.map(processDisplayFieldValues).join(', ')
+  } else if (typeof value === 'number') {
+    return value.toString()
+  } else if (typeof value === 'string') {
+    return value
+  } else if (typeof value === 'object' && value !== null) {
+    const processedObject = {}
+    for (const key in value) {
+      processedObject[key] = processDisplayFieldValues(value[key])
+    }
+    return processedObject
+  } else {
+    return 'Unsupported Type'
+  }
+}